Commit final

+++ lib/rn/commands/notes.rb
@@ -2,6 +2,7 @@ module RN
   module Commands
     module Notes
       class Create < Dry::CLI::Command
+        require 'tty-editor'
         desc 'Create a note'
 
         argument :title, required: true, desc: 'Title of the note'
@@ -16,6 +17,7 @@ module RN
         def call(title:, **options)
           book = options[:book]
           Dir.chdir("/home/#{ENV["USER"]}")
+          title=title.delete("/")
           if(!Dir.exists? ".my_rns")
             puts "No existe la carpeta .my_rns, debe crearla y crear un book"
           else
@@ -23,13 +25,25 @@ module RN
             if book
               if(Dir.exists? book)
                 Dir.chdir(book)
-                File.open(title, "w")
+                if(File.exists? title)
+                  puts "La nota ya existe en #{Dir.pwd}"
+                else
+                  File.open(title, "w")
+                  TTY::Editor.open("#{title}")
+                  puts "La nota fue creada en #{Dir.pwd}"
+                end
               else
                 puts "No existe el libro que envió como parametro debe crearlo"
               end
             else
               Dir.chdir("cuaderno_global")
-              File.open(title, "w")              
+              if(File.exists? title)
+                puts "La nota ya existe en #{Dir.pwd}"
+              else
+                File.open(title, "w")
+                TTY::Editor.open("#{title}")
+                puts "La nota fue creada en #{Dir.pwd}"
+              end
             end
           end
         end
@@ -57,17 +71,23 @@ module RN
             if book
               if(Dir.exists? book)
                 Dir.chdir(book)
-                if(Dir.exists? title)
+                if(File.exists? title)
                   File.delete(title)
+                  puts "Se eliminó la nota #{title} en el libro #{book}"
                 else
-                  puts "No existe la nota que envio en el libro que envio"
+                  puts "No existe el nombre de la nota "
                 end
               else
                 puts "No existe el libro que envió como parametro debe crearlo"
               end
             else
               Dir.chdir("cuaderno_global")
-              File.delete(title)
+              if(File.exists? title)
+                File.delete(title)
+                puts "Se eliminó la nota #{title} en cuaderno_global"
+              else
+                puts "No existe el nombre de la nota en el cuaderno_global"
+              end
             end
           end
         end
@@ -100,7 +120,7 @@ module RN
                 if(File.exists? title)
                   TTY::Editor.open("#{title}")
                 else
-                  "No existe la nota que envio en el libro que envio"
+                  puts "No existe el nombre de la nota #{title} en el libro #{book}"
                 end
               else
                 puts "No existe el libro que envió como parametro debe crearlo"
@@ -110,7 +130,7 @@ module RN
               if(File.exists? title)
                 TTY::Editor.open("#{title}")
               else
-                "No existe la nota que envio en el libro que envio"
+                puts "No existe el nombre de la nota en el cuaderno_global"
               end
             end
           end
@@ -143,7 +163,7 @@ module RN
                 if(File.exists? old_title)
                   File.rename(old_title, new_title)
                 else
-                  "No existe la nota que envio en el libro que envio"
+                  puts "No existe el nombre de la nota #{title} en el libro #{book}"
                 end
               else
                 puts "No existe el libro que envió como parametro debe crearlo"
@@ -153,7 +173,7 @@ module RN
               if(File.exists? old_title)
                 File.rename(old_title, new_title)
               else
-                "No existe la nota que envio en el libro que envio"
+                puts "No existe el nombre de la nota en el cuaderno_global"
               end
             end
           end
@@ -221,7 +241,7 @@ module RN
                 if(File.exists? title)
                   puts File.read(title)
                 else
-                  "No existe la nota que envio en el libro que envio"
+                  puts "No existe el nombre de la nota #{title} en el libro #{book}"
                 end
               else
                 puts "No existe el libro que envió como parametro debe crearlo"
@@ -231,7 +251,7 @@ module RN
               if(File.exists? title)
                 puts File.read(title)
               else
-                "No existe la nota que envio en el libro que envio"
+                puts "No existe el nombre de la nota en el cuaderno_global"
               end
             end
           end
